--- a/scalpel-core/src/Text/HTML/Scalpel/Internal/Scrape.hs
+++ b/scalpel-core/src/Text/HTML/Scalpel/Internal/Scrape.hs
@@ -39,50 +39,49 @@ import Control.Monad.Writer (MonadWriter)
 import Control.Monad.Fix
 import Data.Functor.Identity
 import Data.Maybe
+import Data.Text (Text)
 
 import qualified Control.Monad.Fail as Fail
 import qualified Data.Vector as Vector
-import qualified Text.HTML.TagSoup as TagSoup
-import qualified Text.StringLike as TagSoup
-
+import qualified Data.Text as T
+import qualified Data.Text.Lazy as TL
+import qualified Text.HTML.Parser as HP
 
 -- | A 'ScraperT' operates like 'Scraper' but also acts as a monad transformer.
-newtype ScraperT str m a = MkScraper (ReaderT (TagSpec str) (MaybeT m) a)
+newtype ScraperT m a = MkScraper (ReaderT (TagSpec) (MaybeT m) a)
     deriving (Functor, Applicative, Alternative, Monad, MonadPlus, MonadFix,
               MonadIO, MonadCont, MonadError e, MonadState s, MonadWriter w)
 
 #if MIN_VERSION_base(4,9,0)
-deriving instance Monad m => Fail.MonadFail (ScraperT str m)
+deriving instance Monad m => Fail.MonadFail (ScraperT m)
 #else
-instance Fail.MonadFail m => Fail.MonadFail (ScraperT str m) where
+instance Fail.MonadFail m => Fail.MonadFail (ScraperT m) where
   fail = lift . Fail.fail
 #endif
 
-instance MonadTrans (ScraperT str) where
+instance MonadTrans (ScraperT) where
   lift op = MkScraper . lift . lift $ op
 
-instance MonadReader s m => MonadReader s (ScraperT str m) where
+instance MonadReader s m => MonadReader s (ScraperT m) where
   ask = MkScraper (lift . lift $ ask)
   local f (MkScraper op) = (fmap MkScraper . mapReaderT . local) f op
 
 -- | A value of 'Scraper' @a@ defines a web scraper that is capable of consuming
--- a list of 'TagSoup.Tag's and optionally producing a value of type @a@.
-type Scraper str = ScraperT str Identity
+-- a list of 'HP.Tag's and optionally producing a value of type @a@.
+type Scraper = ScraperT Identity
 
-scrapeTagSpec :: ScraperT str m a -> TagSpec str -> m (Maybe a)
+scrapeTagSpec :: ScraperT m a -> TagSpec -> m (Maybe a)
 scrapeTagSpec (MkScraper r) = runMaybeT . runReaderT r
 
--- | The 'scrapeT' function executes a 'ScraperT' on a list of 'TagSoup.Tag's
+-- | The 'scrapeT' function executes a 'ScraperT' on a list of 'HP.Tag's
 -- and produces an optional value. Since 'ScraperT' is a monad transformer, the
 -- result is monadic.
-scrapeT :: (TagSoup.StringLike str)
-        => ScraperT str m a -> [TagSoup.Tag str] -> m (Maybe a)
-scrapeT s = scrapeTagSpec s . tagsToSpec . TagSoup.canonicalizeTags
+scrapeT :: ScraperT m a -> [HP.Token] -> m (Maybe a)
+scrapeT s = scrapeTagSpec s . tagsToSpec . HP.canonicalizeTokens
 
--- | The 'scrape' function executes a 'Scraper' on a list of 'TagSoup.Tag's and
+-- | The 'scrape' function executes a 'Scraper' on a list of 'HP.Tag's and
 -- produces an optional value.
-scrape :: (TagSoup.StringLike str)
-       => Scraper str a -> [TagSoup.Tag str] -> Maybe a
+scrape :: Scraper a -> [HP.Token] -> Maybe a
 scrape = fmap runIdentity . scrapeT
 
 -- | The 'chroot' function takes a selector and an inner scraper and executes
@@ -91,8 +90,8 @@ scrape = fmap runIdentity . scrapeT
 --
 -- This function will match only the first set of tags matching the selector, to
 -- match every set of tags, use 'chroots'.
-chroot :: (TagSoup.StringLike str, Monad m)
-       => Selector -> ScraperT str m a -> ScraperT str m a
+chroot :: (Monad m)
+       => Selector -> ScraperT m a -> ScraperT m a
 chroot selector inner = do
     maybeResult <- listToMaybe <$> chroots selector inner
     guard (isJust maybeResult)
@@ -105,8 +104,8 @@ chroot selector inner = do
 --
 -- > s = "<div><div>A</div></div>"
 -- > scrapeStringLike s (chroots "div" (pure 0)) == Just [0, 0]
-chroots :: (TagSoup.StringLike str, Monad m)
-        => Selector -> ScraperT str m a -> ScraperT str m [a]
+chroots :: (Monad m)
+        => Selector -> ScraperT m a -> ScraperT m [a]
 chroots selector (MkScraper (ReaderT inner)) =
         MkScraper $ ReaderT $ \tags -> MaybeT $ do
           mvalues <- forM (select selector tags) (runMaybeT . inner)
@@ -114,7 +113,7 @@ chroots selector (MkScraper (ReaderT inner)) =
 
 -- | The 'matches' function takes a selector and returns `()` if the selector
 -- matches any node in the DOM.
-matches :: (TagSoup.StringLike str, Monad m) => Selector -> ScraperT str m ()
+matches :: (Monad m) => Selector -> ScraperT m ()
 matches s = MkScraper $ (guard . not . null) =<< reader (select s)
 
 -- | The 'text' function takes a selector and returns the inner text from the
@@ -122,7 +121,7 @@ matches s = MkScraper $ (guard . not . null) =<< reader (select s)
 --
 -- This function will match only the first set of tags matching the selector, to
 -- match every set of tags, use 'texts'.
-text :: (TagSoup.StringLike str, Monad m) => Selector -> ScraperT str m str
+text :: (Monad m) => Selector -> ScraperT m Text
 text s = MkScraper $ withHead tagsToText =<< reader (select s)
 
 -- | The 'texts' function takes a selector and returns the inner text from every
@@ -130,8 +129,8 @@ text s = MkScraper $ withHead tagsToText =<< reader (select s)
 --
 -- > s = "<div>Hello <div>World</div></div>"
 -- > scrapeStringLike s (texts "div") == Just ["Hello World", "World"]
-texts :: (TagSoup.StringLike str, Monad m)
-      => Selector -> ScraperT str m [str]
+texts :: (Monad m)
+      => Selector -> ScraperT m [Text]
 texts s = MkScraper $ withAll tagsToText =<< reader (select s)
 
 -- | The 'html' function takes a selector and returns the html string from the
@@ -139,7 +138,7 @@ texts s = MkScraper $ withAll tagsToText =<< reader (select s)
 --
 -- This function will match only the first set of tags matching the selector, to
 -- match every set of tags, use 'htmls'.
-html :: (TagSoup.StringLike str, Monad m) => Selector -> ScraperT str m str
+html :: (Monad m) => Selector -> ScraperT m Text
 html s = MkScraper $ withHead tagsToHTML =<< reader (select s)
 
 -- | The 'htmls' function takes a selector and returns the html string from
@@ -147,8 +146,8 @@ html s = MkScraper $ withHead tagsToHTML =<< reader (select s)
 --
 -- > s = "<div><div>A</div></div>"
 -- > scrapeStringLike s (htmls "div") == Just ["<div><div>A</div></div>", "<div>A</div>"]
-htmls :: (TagSoup.StringLike str, Monad m)
-      => Selector -> ScraperT str m [str]
+htmls :: (Monad m)
+      => Selector -> ScraperT m [Text]
 htmls s = MkScraper $ withAll tagsToHTML =<< reader (select s)
 
 -- | The 'innerHTML' function takes a selector and returns the inner html string
@@ -157,8 +156,8 @@ htmls s = MkScraper $ withAll tagsToHTML =<< reader (select s)
 --
 -- This function will match only the first set of tags matching the selector, to
 -- match every set of tags, use 'innerHTMLs'.
-innerHTML :: (TagSoup.StringLike str, Monad m)
-          => Selector -> ScraperT str m str
+innerHTML :: (Monad m)
+          => Selector -> ScraperT m Text
 innerHTML s = MkScraper $ withHead tagsToInnerHTML =<< reader (select s)
 
 -- | The 'innerHTMLs' function takes a selector and returns the inner html
@@ -166,8 +165,8 @@ innerHTML s = MkScraper $ withHead tagsToInnerHTML =<< reader (select s)
 --
 -- > s = "<div><div>A</div></div>"
 -- > scrapeStringLike s (innerHTMLs "div") == Just ["<div>A</div>", "A"]
-innerHTMLs :: (TagSoup.StringLike str, Monad m)
-           => Selector -> ScraperT str m [str]
+innerHTMLs :: (Monad m)
+           => Selector -> ScraperT m [Text]
 innerHTMLs s = MkScraper $ withAll tagsToInnerHTML =<< reader (select s)
 
 -- | The 'attr' function takes an attribute name and a selector and returns the
@@ -176,11 +175,11 @@ innerHTMLs s = MkScraper $ withAll tagsToInnerHTML =<< reader (select s)
 --
 -- This function will match only the opening tag matching the selector, to match
 -- every tag, use 'attrs'.
-attr :: (Show str, TagSoup.StringLike str, Monad m)
-     => String -> Selector -> ScraperT str m str
+attr :: (Monad m)
+     => Text -> Selector -> ScraperT m Text
 attr name s = MkScraper $ ReaderT $ MaybeT
               . return . listToMaybe . catMaybes
-              . fmap (tagsToAttr $ TagSoup.castString name) . select s
+              . fmap (tagsToAttr name) . select s
 
 -- | The 'attrs' function takes an attribute name and a selector and returns the
 -- value of the attribute of the given name for every opening tag
@@ -188,12 +187,11 @@ attr name s = MkScraper $ ReaderT $ MaybeT
 --
 -- > s = "<div id=\"out\"><div id=\"in\"></div></div>"
 -- > scrapeStringLike s (attrs "id" "div") == Just ["out", "in"]
-attrs :: (Show str, TagSoup.StringLike str, Monad m)
-     => String -> Selector -> ScraperT str m [str]
+attrs :: (Monad m)
+     => Text -> Selector -> ScraperT m [Text]
 attrs name s = MkScraper $ ReaderT $ MaybeT
                . return . Just . catMaybes
-               . fmap (tagsToAttr nameStr) . select s
-    where nameStr = TagSoup.castString name
+               . fmap (tagsToAttr name) . select s
 
 -- | The 'position' function is intended to be used within the do-block of a
 -- `chroots` call. Within the do-block position will return the index of the
@@ -229,43 +227,41 @@ attrs name s = MkScraper $ ReaderT $ MaybeT
 -- , (2, "Third paragraph.")
 -- ]
 -- @
-position :: (TagSoup.StringLike str, Monad m) => ScraperT str m Int
+position :: (Monad m) => ScraperT m Int
 position = MkScraper $ reader tagsToPosition
 
-withHead :: Monad m => (a -> b) -> [a] -> ReaderT (TagSpec str) (MaybeT m) b
+withHead :: Monad m => (a -> b) -> [a] -> ReaderT TagSpec (MaybeT m) b
 withHead _ []    = empty
 withHead f (x:_) = return $ f x
 
-withAll :: Monad m => (a -> b) -> [a] -> ReaderT (TagSpec str) (MaybeT m) [b]
+withAll :: Monad m => (a -> b) -> [a] -> ReaderT TagSpec (MaybeT m) [b]
 withAll f xs = return $ map f xs
 
-foldSpec :: TagSoup.StringLike str
-         => (TagSoup.Tag str -> str -> str) -> TagSpec str -> str
-foldSpec f = Vector.foldr' (f . infoTag) TagSoup.empty . (\(a, _, _) -> a)
+foldSpec ::(HP.Token -> Text -> Text) -> TagSpec -> Text
+foldSpec f = Vector.foldr' (f . infoTag) T.empty . (\(a, _, _) -> a)
 
 
-tagsToText :: TagSoup.StringLike str => TagSpec str -> str
+tagsToText :: TagSpec -> Text
 tagsToText = foldSpec f
     where
-        f (TagSoup.TagText str) s = str `TagSoup.append` s
+        f (HP.ContentText t) s = t `T.append` s
         f _                     s = s
 
-tagsToHTML :: TagSoup.StringLike str => TagSpec str -> str
-tagsToHTML = foldSpec (\tag s -> TagSoup.renderTags [tag] `TagSoup.append` s)
+tagsToHTML :: TagSpec -> Text
+tagsToHTML = foldSpec (\tag s -> (TL.toStrict $ HP.renderTokens [tag]) `T.append` s)
 
-tagsToInnerHTML :: TagSoup.StringLike str => TagSpec str -> str
+tagsToInnerHTML :: TagSpec -> Text
 tagsToInnerHTML (tags, tree, ctx)
-    | len < 2   = TagSoup.empty
+    | len < 2   = T.empty
     | otherwise = tagsToHTML (Vector.slice 1 (len - 2) tags, tree, ctx)
     where len = Vector.length tags
 
-tagsToAttr :: (Show str, TagSoup.StringLike str)
-           => str -> TagSpec str -> Maybe str
+tagsToAttr :: Text -> TagSpec -> Maybe Text
 tagsToAttr tagName (tags, _, _) = do
     guard $ 0 < Vector.length tags
     let tag = infoTag $ tags Vector.! 0
-    guard $ TagSoup.isTagOpen tag
-    return $ TagSoup.fromAttrib tagName tag
+    guard $ isTagOpen tag
+    return $ fromAttrib tagName tag
 
-tagsToPosition :: TagSpec str -> Int
+tagsToPosition :: TagSpec -> Int
 tagsToPosition (_, _, ctx) = ctxPosition ctx