--- a/scalpel-core/src/Text/HTML/Scalpel/Internal/Select/Types.hs
+++ b/scalpel-core/src/Text/HTML/Scalpel/Internal/Select/Types.hs
@@ -65,7 +65,7 @@ data SelectSettings = SelectSettings {
   -- | The required depth of the current select node in relation to the
   -- previously matched SelectNode.
   selectSettingsDepth :: Maybe Int
-}
+} deriving Show
 
 defaultSelectSettings :: SelectSettings
 defaultSelectSettings = SelectSettings {