--- a/scalpel/scalpel.cabal
+++ b/scalpel/scalpel.cabal
@@ -38,7 +38,7 @@ library
   default-language: Haskell2010
   build-depends:
           base            >= 4.6 && < 5
-      ,   scalpel-core    == 0.6.2.2
+      ,   scalpel-core
       ,   bytestring
       ,   case-insensitive
       ,   data-default