… README.md
… scalpel-core/benchmarks/Main.hs
… scalpel-core/src/Text/HTML/Scalpel/Internal/Serial.hs
… scalpel-core/tests/TestMain.hs
… scalpel/.ghc.environment.aarch64-darwin-9.12.2
… scalpel/src/Text/HTML/Scalpel.hs
… scalpel/src/Text/HTML/Scalpel/Internal/Scrape/URL.hs