solves #109

## Effect
### practical example
_see https://github.com/benjaminweb/scraper-experiment_

task: scraping all _h2_ tags on amazon.com html
before: peak memory 130 MB, runtime: 160 ms
after: peak memory allocation: 13 MB, runtime 11 ms
Rust: peak memory allocation: 3 MB, runtime: 4.7 ms

=> this change puts scalpel almost into the same ballpark Rust is in!

_(on first sight, the packaged benchmarks seem not to reflect this good – might it be valuable to include some real world example?)_

### builtin benchmarks

Before:
```
Benchmark bench: RUNNING...
benchmarking nested/100
time                 48.45 μs   (48.20 μs .. 48.80 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 48.70 μs   (48.52 μs .. 48.94 μs)
std dev              674.7 ns   (499.8 ns .. 1.019 μs)

benchmarking nested/1000
time                 632.0 μs   (630.4 μs .. 634.4 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 634.2 μs   (631.5 μs .. 638.8 μs)
std dev              11.62 μs   (7.512 μs .. 16.75 μs)

benchmarking nested/10000
time                 14.36 ms   (13.97 ms .. 14.76 ms)
                     0.995 R²   (0.991 R² .. 0.999 R²)
mean                 13.75 ms   (13.57 ms .. 13.97 ms)
std dev              525.8 μs   (381.1 μs .. 793.9 μs)
variance introduced by outliers: 14% (moderately inflated)

benchmarking many-selects/10
time                 118.9 ns   (118.4 ns .. 119.7 ns)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 119.9 ns   (119.1 ns .. 121.4 ns)
std dev              3.576 ns   (2.399 ns .. 6.277 ns)
variance introduced by outliers: 46% (moderately inflated)

benchmarking many-selects/100
time                 1.089 μs   (1.083 μs .. 1.098 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 1.098 μs   (1.092 μs .. 1.106 μs)
std dev              24.33 ns   (18.40 ns .. 32.31 ns)
variance introduced by outliers: 27% (moderately inflated)

benchmarking many-selects/1000
time                 10.78 μs   (10.72 μs .. 10.85 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 10.87 μs   (10.81 μs .. 10.97 μs)
std dev              244.8 ns   (172.9 ns .. 362.6 ns)
variance introduced by outliers: 23% (moderately inflated)

benchmarking many-///10
time                 429.5 μs   (422.2 μs .. 435.7 μs)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 421.7 μs   (419.3 μs .. 425.1 μs)
std dev              9.358 μs   (7.277 μs .. 11.82 μs)
variance introduced by outliers: 14% (moderately inflated)

benchmarking many-///100
time                 437.4 μs   (431.4 μs .. 446.8 μs)
                     0.998 R²   (0.996 R² .. 0.999 R²)
mean                 439.9 μs   (435.9 μs .. 444.6 μs)
std dev              14.78 μs   (12.65 μs .. 19.08 μs)
variance introduced by outliers: 26% (moderately inflated)

benchmarking many-///1000
time                 592.6 μs   (587.4 μs .. 599.3 μs)
                     0.999 R²   (0.997 R² .. 0.999 R²)
mean                 594.2 μs   (588.8 μs .. 601.5 μs)
std dev              20.68 μs   (15.15 μs .. 28.26 μs)
variance introduced by outliers: 27% (moderately inflated)

Benchmark bench: FINISH
```


After:
```
Benchmark bench: RUNNING...
benchmarking nested/100
time                 25.80 μs   (25.61 μs .. 26.01 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 26.00 μs   (25.82 μs .. 26.23 μs)
std dev              688.5 ns   (492.0 ns .. 897.3 ns)
variance introduced by outliers: 27% (moderately inflated)

benchmarking nested/1000
time                 300.9 μs   (300.0 μs .. 301.7 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 303.2 μs   (302.3 μs .. 305.6 μs)
std dev              4.595 μs   (2.197 μs .. 8.768 μs)

benchmarking nested/10000
time                 5.946 ms   (5.907 ms .. 5.996 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 5.831 ms   (5.797 ms .. 5.865 ms)
std dev              98.94 μs   (82.26 μs .. 126.2 μs)

benchmarking many-selects/10
time                 114.5 ns   (114.2 ns .. 114.9 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 114.7 ns   (114.5 ns .. 115.1 ns)
std dev              982.6 ps   (670.3 ps .. 1.472 ns)

benchmarking many-selects/100
time                 1.053 μs   (1.050 μs .. 1.055 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.053 μs   (1.050 μs .. 1.055 μs)
std dev              7.886 ns   (6.514 ns .. 10.16 ns)

benchmarking many-selects/1000
time                 10.46 μs   (10.41 μs .. 10.51 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 10.47 μs   (10.44 μs .. 10.50 μs)
std dev              102.8 ns   (86.66 ns .. 124.9 ns)

benchmarking many-///10
time                 150.7 μs   (149.3 μs .. 152.1 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 150.1 μs   (149.5 μs .. 150.8 μs)
std dev              2.215 μs   (1.731 μs .. 2.890 μs)

benchmarking many-///100
time                 156.8 μs   (156.2 μs .. 157.6 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 157.5 μs   (157.0 μs .. 158.2 μs)
std dev              1.961 μs   (1.179 μs .. 3.021 μs)

benchmarking many-///1000
time                 253.1 μs   (252.6 μs .. 253.6 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 250.0 μs   (248.9 μs .. 250.8 μs)
std dev              3.134 μs   (2.641 μs .. 3.924 μs)

Benchmark bench: FINISH
```

## 8 broken tests require fixing

@fimad I've tried to spot the cause(s) of it to no avail yet – can you pinpoint them?

```
Running 1 test suites...
Test suite lib-tests: RUNNING...
### Failure in: 0:scrapeTests:22:"attr extract matching attribute value: scrape (<a key=foo />)"
tests/TestMain.hs:626
expected: Just "foo"
 but got: Nothing
### Failure in: 0:scrapeTests:23:"attr extract matching attribute value with complex predicates: scrape (<a key1=foo/><b key1=bar key2=foo /><a key1=bar key2=baz />)"
tests/TestMain.hs:626
expected: Just "baz"
 but got: Nothing
### Failure in: 0:scrapeTests:29:"scraping should work for self-closing tags: scrape (<img src='foobar' />)"
tests/TestMain.hs:626
expected: Just "foobar"
 but got: Nothing
### Failure in: 0:scrapeTests:30:"lower case selectors should match any case tag: scrape (<a>foo</a><A>bar</A>)"
tests/TestMain.hs:626
expected: Just ["foo","bar"]
 but got: Just ["foo"]
### Failure in: 0:scrapeTests:31:"upper case selectors should match any case tag: scrape (<a>foo</a><A>bar</A>)"
tests/TestMain.hs:626
expected: Just ["foo","bar"]
 but got: Just ["foo"]
### Failure in: 0:scrapeTests:32:"attribute key matching should be case-insensitive: scrape (<a B=C>foo</a>)"
tests/TestMain.hs:626
expected: Just ["foo"]
 but got: Just []
### Failure in: 0:scrapeTests:34:"notP should invert attribute value matching: scrape (<a>foo</a><a B=C>bar</a><a B=D>baz</a>)"
tests/TestMain.hs:626
expected: Just ["foo","baz"]
 but got: Just ["foo","bar","baz"]
### Failure in: 0:scrapeTests:82:"untilBack should leave the focus of the new context at the end: scrape (<b foo=bar /><a>1</a><a>2</a><a>3</a>)"
tests/TestMain.hs:626
expected: Just ("bar",["1","2","3"],["2","1"])
 but got: Nothing
Cases: 86  Tried: 86  Errors: 0  Failures: 8
Test suite lib-tests: FAIL
```