@benjaminweb 2025-05-22
Update: fixed 4 broken tests, 4 remaining (which relate all to case-sensitive variants):

```### Failure in: 0:scrapeTests:30:"lower case selectors should match any case tag: scrape (<a>foo</a><A>bar</A>)"
tests/TestMain.hs:626
expected: Just ["foo","bar"]
 but got: Just ["foo"]
### Failure in: 0:scrapeTests:31:"upper case selectors should match any case tag: scrape (<a>foo</a><A>bar</A>)"
tests/TestMain.hs:626
expected: Just ["foo","bar"]
 but got: Just ["bar"]
### Failure in: 0:scrapeTests:32:"attribute key matching should be case-insensitive: scrape (<a B=C>foo</a>)"
tests/TestMain.hs:626
expected: Just ["foo"]
 but got: Just []
### Failure in: 0:scrapeTests:34:"notP should invert attribute value matching: scrape (<a>foo</a><a B=C>bar</a><a B=D>baz</a>)"
tests/TestMain.hs:626
expected: Just ["foo","baz"]
 but got: Just ["foo","bar","baz"]
Cases: 86  Tried: 86  Errors: 0  Failures: 4```