@fimad 2025-05-25
This is awesome, I really appreciate you taking this on!

I looked through your changes and don't have any major feedback. I also checked out `html-parse`, the library looks solid.

Your observation about the currently packaged benchmarks is accurate. They were originally written to trigger some pathological cases when I was using a way inferior algorithm. A little while ago I played around with some local changes that tried to speed things up and realized that these benchmarks aren't really helpful for understanding real world use cases.

If you want to upstream your more realistic benchmarks I'd be happy to include those as well. Though maybe swap to a Wikipedia article since I think those would be licensed under CC. I'm not sure about checking in other sites HTML.